refactor(dispatch): split provider_loop 808 LoC into resolver + retry + orchestration (T-SP-3 #9)
provider_loop.rs was hosting three distinct concerns in a single file:
1. Provider selection (registry + CB + HC + RE-1a endpoint health)
2. Retry + error classification + single-provider dispatch
3. Outer fallback loop orchestration

Split into:
- src/server/dispatch/resolver.rs (~128 LoC) — resolve_provider,
  try_direct_provider_lookup (backward-compat path).
- src/server/dispatch/retry.rs (~482 LoC) — ProviderLoopAction,
  ProviderAttempt, classify_and_handle_error, emit_provider_error_metrics,
  handle_provider_error, try_rotate_and_retry, dispatch_streaming,
  dispatch_non_streaming, wrap_stream_with_middleware, build_encrypted_content.
- src/server/dispatch/provider_loop.rs (~275 LoC down from 808) —
  dispatch_provider_loop (outer loop), log_dispatch_attempt,
  prepare_provider_request, emit_fallback.

No behavior change: public entry point src/server/dispatch/mod.rs:349
(provider_loop::dispatch_provider_loop) keeps the same signature and
callers are unaffected.

Validation:
- cargo build --all-targets green
- cargo test --lib : 949 passed, 0 failed
- cargo clippy --all-targets -- -D warnings clean
- RUSTFLAGS="-D warnings" cargo check --no-default-features clean

Addresses Phoenix audit 2026-04-20 item #9.
